refactor(scheduled-tasks): dedup formatRunTimestamp into a shared util
The runs timestamp formatter was byte-identical in schedule-runs-list.tsx and
schedule-triggers-client.tsx. Extract to lib/utils/format-run-timestamp.ts with
a unit test; both callers import it. Behavior-preserving (/unify Stage 0).
