+++ platform/frontend/src/app/scheduled-tasks/schedule-triggers-client.tsx
@@ -75,6 +75,7 @@ import { useMyTeams } from "@/lib/teams/team.query";
 import { cn } from "@/lib/utils";
 import { formatRelativeTimeFromNow } from "@/lib/utils/date-time";
 import { formatCronSchedule } from "@/lib/utils/format-cron";
+import { formatRunTimestamp } from "@/lib/utils/format-run-timestamp";
 import {
   type AgentOption,
   buildCronFromSchedule,
@@ -955,43 +956,6 @@ function getDefaultTriggerName(
 ): string {
   const agent = agentOptions.find((a) => a.value === agentId);
   return agent ? `Scheduled ${agent.label}` : "";
-}
-
-function formatRunTimestamp(dateString: string): string {
-  const date = new Date(dateString);
-  const now = new Date();
-
-  const timeStr = date.toLocaleTimeString(undefined, {
-    hour: "numeric",
-    minute: "2-digit",
-  });
-
-  const isToday =
-    date.getFullYear() === now.getFullYear() &&
-    date.getMonth() === now.getMonth() &&
-    date.getDate() === now.getDate();
-
-  if (isToday) {
-    return `Today at ${timeStr}`;
-  }
-
-  const yesterday = new Date(now);
-  yesterday.setDate(yesterday.getDate() - 1);
-  const isYesterday =
-    date.getFullYear() === yesterday.getFullYear() &&
-    date.getMonth() === yesterday.getMonth() &&
-    date.getDate() === yesterday.getDate();
-
-  if (isYesterday) {
-    return `Yesterday at ${timeStr}`;
-  }
-
-  const dateStr = date.toLocaleDateString(undefined, {
-    month: "short",
-    day: "numeric",
-  });
-
-  return `${dateStr} at ${timeStr}`;
 }
 
 function RunStatusIcon({ status }: { status: ScheduleTriggerRunStatus }) {
