+++ platform/frontend/src/components/scheduled-tasks/schedule-runs-list.tsx
@@ -15,6 +15,7 @@ import {
   useScheduleTriggerRuns,
 } from "@/lib/schedule-trigger.query";
 import { cn } from "@/lib/utils";
+import { formatRunTimestamp } from "@/lib/utils/format-run-timestamp";
 
 /**
  * A schedule's runs, reused by the project runs page and the chat right-side
@@ -141,41 +142,4 @@ function RunRow({
   return (
     <div className="rounded-lg border bg-card opacity-80">{rowContent}</div>
   );
-}
-
-function formatRunTimestamp(dateString: string): string {
-  const date = new Date(dateString);
-  const now = new Date();
-
-  const timeStr = date.toLocaleTimeString(undefined, {
-    hour: "numeric",
-    minute: "2-digit",
-  });
-
-  const isToday =
-    date.getFullYear() === now.getFullYear() &&
-    date.getMonth() === now.getMonth() &&
-    date.getDate() === now.getDate();
-
-  if (isToday) {
-    return `Today at ${timeStr}`;
-  }
-
-  const yesterday = new Date(now);
-  yesterday.setDate(yesterday.getDate() - 1);
-  const isYesterday =
-    date.getFullYear() === yesterday.getFullYear() &&
-    date.getMonth() === yesterday.getMonth() &&
-    date.getDate() === yesterday.getDate();
-
-  if (isYesterday) {
-    return `Yesterday at ${timeStr}`;
-  }
-
-  const dateStr = date.toLocaleDateString(undefined, {
-    month: "short",
-    day: "numeric",
-  });
-
-  return `${dateStr} at ${timeStr}`;
 }
